refactor: moved tests into global tests folder
